Move method to utility module
